feat: gameService 와 도메인 연결
