+++ src/main/java/bowling/GameService.java
@@ -1,0 +1,47 @@
+package bowling;
+
+
+import bowling.domain.Frame;
+import bowling.domain.NormalFrame;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class GameService {
+    private final List<Frame> frames = new ArrayList<>();
+
+    public GameService(String playerName) {
+        if (playerName.length() > 3) {
+            throw new IllegalPlayerNameException();
+        }
+    }
+
+    public boolean isFinish() {
+        return getCurrentFrame().isFinish();
+    }
+
+    public int getCurrentFrameNumber() {
+        return getCurrentFrame().getNumber();
+    }
+
+    private Frame getCurrentFrame() {
+        return frames.get(frames.size() - 1);
+    }
+
+    public GameStatus throwBall(int fallingPins) {
+        Frame frame = getCurrentFrame().throwBall(fallingPins);
+        updateFrames(frame);
+        return new GameStatus(this);
+    }
+
+    private void updateFrames(Frame frame) {
+        if (getCurrentFrame() != frame) {
+            frames.add(frame);
+        }
+    }
+
+    public GameStatus start() {
+        frames.add(new NormalFrame());
+        return new GameStatus(this);
+    }
+}
