+++ src/main/java/bowling/GameStatus.java
@@ -1,4 +1,7 @@
 package bowling;
 
 public class GameStatus {
+    public GameStatus(GameService game) {
+
+    }
 }
